Enhanced navigation in fullscreen mode.
Adjusted the video selection/mounting and navigation (prev/next) to work complete on the li rather than on the text input.

Also added a keydown handler that checks to see if the text input is fully selected, and if so, allow the "f" and space keys to run their functions (fullscreen toggle and video play toggle), but only if the input is fully selected.  This will help avoid people accidentally erasing a playlist item with a space or "f".