Video objects now stored in li element directly.
This eliminates the weirdness of a separate map variable that tracked the li elements but in an indirect way that would also memory leak.

Also paves the way for multiple video objects to be associated with a single playlist item. This is needed for two reasons: Playlist items that have multiple scriptures from different Bible Books separated by semicolons, and incomplete Bible Books that have individual verse video files. So a `videos` property is associated with the li element and is an array of video objects.

Right now a few spots in the code expect to look for position [0] but in the future they should be able to handle multiple video objects.