Support for multiple videos for a playlist item.
So when playing a video, we keep track of the video index and cue index, with video index being which video in the array of videos, and cue index being which cue in the list of cues a given video may have.

This also meant the calculation of percent completed had to be rewritten and couldn't fit within the ScriptureVideo class itself, because now the calculation spans multiple video objects. This leads me to think that a playlist item class would be beneficial.

One more thing.. note that when parsing the text into video objects, I couldn't just push the video objects into an array because the video object creation happens asynchronously. So we create the array ahead of time and put the created objects in the right spot in the array as they come in.