Finally, support for verse videos!
Bible books that haven't been completely translated are stored in JW app as single videos for each verse.

This was challenging because there was no way of handling one scripture, which could be a range of verses, when that may require multiple video files. It was always expected that ONE scripture was synonymous with ONE video file.

Work was done on a previous commit to handle multiple videos for a single playlist item. This also helps handle playing multiple scriptures in a single playlist item like "Gen 1:1; Ex 2:3". But now this commit builds on that to handle single verse videos.

ScriptureVideoUtil always tries chapter files first and then verse files if the chapter cannot be found. If then creates ScriptureVideo objects for each verse in the scripture that needs to use verse videos. It returns them all as an array.

Then index.js handles the storage and display of these "composite scriptures" accordingly.