Add initial support for rewind/fast-forward btns.
Typical rewind 5 sec and fast-forward 15 sec buttons. However, they don't know how to properly navigate the video/cue lists when passing start/stop points, so it is a little wonky. So the feature is added but the buttons are hidden for now. Fine-tune this.

One benefit that users WILL see on this commit, though, is that once a video dismounts (meaning it finishes all videos/cues for a playlist row), the scrubbing buttons (so RW/Play/FF) all become disabled. This helps prevent a user from wondering into video areas that he didn't designate in the playlist.