Provide missing cue feedback a little better.
If a reference or scripture object thinks its ok but its video object cannot find a cue in the video file for it, nothing would happen. It would just be a playlist row with an empty cue list.

Now it has been improved to catch the fact that the cue list is empty, and reports it as an error "Cue not found". Ignores this for images, obviously.

Furthermore, I catch the SPACE on playlist item keyup, since that was inadvertently clearing the error status of the row.