Support for browsing for external media.
See #10.  Made a gear that appears on the playlist row, and additional features can be added to the dropdown menu. For now, it only has a "Browse external media..." option.

The "Browse external media..." option will open a file dialog and let the user select an image or video file. This way the user doesn't have to type the media path into the row.