Added a "Batch Entry…" option. See #8.
Uses jQuery dialog to open a dialog box that allows freeform text entry, and it parses out scriptures and references from the content. When you click "Add Items", any discovered items are inserted into the playlist.

To accomplish this, jQuery UI was added.

Also, playlist row adding was updated to allow either appending or prepending a playlist row. Thus support for the "Insert" of playlist rows.