Auto-install ffprobe to be built into the binary. Fixes #1.
All media util classes are tweaked to optionally receive a video app controller object.  The main script then generates a video app controller that points to bin/ffprobe and passes it into the media utility objects.

The installation of ffprobe occurs with the ffbinaries library, in postinstall.js which is executed, ahem, post-install.  Thus it is present when running via `npm start` and is included in the build when doing an `npm run build`.