Added a webvttpatches directory for special webvtt files to be included with the app.
See #18. Adding this webvttpatches directory enables webvtt files to be included with the release. They take precedence in the webvtt search, thus overriding any typically generated webvtt file. This enables us to override any faulty errors in chapter definitions of video files.

Definitions were then included for "sn 3", which has faulty chapters when pulled in by ffprobe.

To accomplish this, the patch directory is prepended to the array of paths to look in.

Instead of using a calcWebvttPath() file, which generates a single prefered path name of a video file's webvtt counterpart, we use calcWebvttRegEx() which helps us easily search for variants of a file's webvtt partner.

Then, we changed the standard so that even external webvtt files will have the full path hash in the filename. Thus, the search can be based on just filename and typically work without finding false positives.