Add a createStudyVideos() method.
It behaves just like the createVideo() method except it returns an ARRAY of ReferenceVideo objects, based on logic to ascertain the study material from an initially passed-in reference.

This will be used when trying to generate study materials playlist. See #6.

The logic is (possibly over) simple.  It uses the question cues to determine what paragraphs should be played. Then also includes art cues.  That's it. Not perfect, but pretty good.