Initial support for second fullscreen display.
Settings screen lets you set whether the second screen opens or not.

The second screen is a very dumb page with very little logic. But it does have the VideoController and it receives pertinent executions to set the state of the video/image/text.

The VideoController on the main screen detects if a second window is registered with it, and if so, it passes along pertinent executions to it.

Due to possible lag time in the creation of the second window, it requests a status update once it is done being created and going full-screen.