Added feature to purge old webvtt files.
See #27. If a webvtt is generated and then a video is replaced with a newer version, the webvtt may be invalid. So on application initialization, it checks all found video files against its index of webvtt files, and deletes any webvtts that are older than the video file they represent.