Added support to build installers.
On Mac, builds a dmg file, on Windows, builds a simple setup.exe file.

Changed to use electron-builder instead of electron-packager for packaging the app.

The electron-builder strongly recommends, and uses by default, asar packaging of your application. In order to be fully compatible with asar packaging of the app, the webvtt package had to be improved to use execFile which is compatible with unpacking asar binaries for execution. Additionally, the wrapper for the webvtt executable must receive the exact filename of the executable to work. So these small changes were made.