Add ability to set the form of webvtt caching.
See #31.

Provided ability to identify cache mode in settings.  The indexVideos() method is adjusted to include initialization of the cache manager so that it can be used to reset the cache manager when its mode changes.

A "Clear Cache" button executes a purgeWebvttFiles() method which clears all the indexed webvtt files on the current cache mode.

These changes require the CACHEMODES constant to include more details such as a friendly name and description, for use on the settings screen.

Default to internal cache.