External media improved to support filename only for known video paths.
See #32. If an external media item is parsed, and it is detected to exist in the index of media in known video paths ("Videos"/"Movies" folder, "Downloads" folder, or other custom folder), the item will be aware of it and display the filename only.

This required some changes:

External Media extensions constants were updated to be more consistent with how the other utility objects behave, and one constant (ALLEXT) handles the combining of all video and image extensions.

External Media Util receives the list of media paths just like Scripture and Reference Video Util objects, following same protocols.

If an ExternalMedia object is told its path is in the index (pathIsIndexed property), the displayName property will return just the filename.

Note these changes required tweaks to the regex so that filenames could be found as well, not just full paths.

Furthermore, logic was improved in the parsing method to search thru the index of media for filenames that match.

Thus, filenames are supported, but completely external paths will still behave as they did previously.