Add ability to save/load playlist with customizations to exact times/cues.
This will help with #26 because it will store exact "start" and "end" times for cues, as well as the "min" and "max" times for those cues, enabling support for trimming a video as well as support for the UI to handle the trimming.

This will also help with #33 because you'll be able to identify specific cues that the current text parsing engine fails to find.

Changes:

- The ScriptureVideo, ReferenceVideo, and ExternalMedia objects were enhanced in numerous ways: (a) The displayName property is not set initially but is its own property that can be set to a custom value. (b) Support to use a Cue object, which handles start/end/content properties like before but now also min/max. (c) Added source, filename, and extension properties. The source is especially useful to uniformly get the source object driving the video object, regardless of type.

The index.js was then modified to support similar "text" based saving/loading of playlist items by default, but if specified, more exact details can be saved to track the customizations to cues. When loading, a video object will be created thru the same createVideo() method just like parsing by text would accomplish, but then the custom cue list is applied after instantiation. That's it!