Add ability to select a cue to play.
User can click a cue in the cue list and jump straight to that cue.

To accomplish this, I enhanced the mountPlaylistItem() function to receive a cueIndex. This was able to simplify the checkVideo() function too, to more simply rely on mountPlaylistItem() when switching cues.

This works closer to #26. In fact, I also added a trim icon but it is hidden for now.

Also added a concept of a "selected" cue list row, behaving similarly to the main playlist "selected" row. This helps user see which row is currently selected and will also enable the trim button to appear on only the selected row.