Added UI for trimming video clips!
This finally closes #26. A trimming dialog box allows the user to set the start/end points with a slider, as well as change the video and cue names.

To avoid confusion, a playlist item that has been tweaked with trimming and renaming has the text made read-only. The user can delete the playlist item but that's it, as far as text-entry goes, once it has been trimmed. This avoids scenario where user tries to edit the text and the field is re-parsed and trims or renames are lost. Once the limitation is accepted, it appears to work well.