Improve "playing" of picture media.
Allow user to "play" images, and the progress bar shows the progress of the picture playing for the designated time constant in the ExternalMedia class.

Multiple images can be played together cleanly, and the progress bar shows you status. It works just as well when mixed with videos.

You can also pause in the middle of displaying pictures, and resume (it will restart the time for the picture it is on, but it will resume at that picture).

This closes #15 where a playlist item with multiple images would start looping thru them immediately. Oops.