Fix keyup handler for handling reparsing better.
Closes #41. In d1fb2ce27a2925a454916f731f31538b716a1dbf, it was set to clear "data-videos-text" on non-control keypresses. This was to fix an issue when typing in a red "comment only" line. Changing the text would remove the red coloring but then if you changed the text to its previous value, it wouldn't attempt to reparse it.

This fix unintentionally killed the purpose of the "data-videos-text" property since it reset it all the time.

This fix will still clear "data-videos-text" for "red" text fields but for those fields only. All parsed fields will not be reset and thus not reparsed.

Thus fixing #41 since deleting a row was unintionally triggering this code on keyup of the delete key when moving to the new playlist row.

This was a doozy.