Add ability to IMPORT external media into playlist.
That is, copy it into the app's internal media folder, which is included with the playlist zip file.

To accomplish this, improved the video watcher so its "add" and "unlink" events call declared functions, so those functions can be called on-the-fly as well.

Then, anytime we put stuff in the internal media folder (when importing a playlist or importing external media), we add it to the indexed videos immediately, so we don't have to wait for it to be "found" by the watcher.

Note, the current implementation cannot add external media to a playlist item that already has some trimmed source="media" content. It will wipe that out. Need to fix.