Improve the display of on-screen text.
The display of text on the screen was poor. It wasn't centered well and it would be weird if it was more than one line.

This fix simplifies the CSS and puts logic to center the text in the VideoController object. It sets the text and then vertically centers the text based on its resulting height and the height of the video container.

The result is great centered text, even for multiple lines, in both playlist, fullscreen, and secondary screen displays!