Add support for date-based pubs in the Publication Lookup! W00T!
Closes #33. Date-based pubs are searched by expected the filename to have a date naming of YYYYMM or YYYYMMDD which they all do presently.

Even displays the issues according to each publication's date formatting style! Since it uses the Publication class's own dateFormatted property, it will even adjust for changes in the Watchtower or Awake! date formatting over the years.

I've noted that a handful of date-based publications (including the latest 2018-03 Awake) do not have cues for the small articles. This is unavoidable and outside the scope of the goal for this task.