Autocheck opening scripture for WT lessons.
See #46. If the reference is a WT and we can find the "Title" cue, the opening scripture is always right after that.  On creation of the list, physically click the checkbox if we can find it.

This is a pretty harmless way to find and click it, that should fail gracefully.