Added basic experimental support for importing SLS playlists.
When unzipping a file, if "playlist.xml" is found, it will treat it like an SLS playlist and convert it to the expected playlist.json format. Uses xml2js library to parse the XML into a JS object.

Currently supported:
- Scriptures, with trimming
- External media, but no trimming
- References, but it will play the entire reference.

So the biggest scenarios that will not work are: An external media video that needs trimming, or any references that are not playing the entire video. Significant lacking features, but this is just an initial experimental implementation, and it will cover a typical public talk scenario with songs, scriptures, and pictures.