Fix study wizard to not have hesitation when including questions.
Previously, questions were just added as a cue to the playlist entry. Since the question was literally immediately after the paragraph cue, there would be a milliseconds-long hesitation as the cue is set.

Now the paragraph cue is adjusted to include the question cue as one single cue. The end/max times are adjusted. So the player just continues playing and there is no hesitation.